between {data.table}	R Documentation

Convenience function for range subset logic.

Description

Intended for use in [.data.table i.

Usage

between(x,lower,upper,incbounds=TRUE)
x 

Arguments

x
Any vector e.g. numeric, character, date, ...

lower
Lower range bound.

upper
Upper range bound.

incbounds
TRUE means inclusive bounds i.e. [lower,upper]. FALSE means exclusive bounds i.e. (lower,upper).

Value

Logical vector as the same length as x with value TRUE for those that lie within the range [lower,upper] or (lower,upper).

Note

Current implementation does not make use of ordered keys.

See Also

data.table, like

Examples

DT = data.table(a=1:5, b=6:10)
DT[b %between% c(7,9)]

[Package data.table version 1.9.6 Index]