first {data.table}	R Documentation

First item of an object

Description

Returns the first item of a vector or list, or the first row of a data.frame or data.table.

Usage

first(x, ...)

Arguments

x
A vector, list, data.frame or data.table. Otherwise the S3 method of xts::first is deployed.

...
Not applicable for data.table::first. Any arguments here are passed through to xts::first.

Value

If no other arguments are supplied it depends on the type of x. The first item of a vector or list. The first row of a data.frame or data.table. Otherwise, whatever xts::first returns (if package xts has been loaded, otherwise a helpful error).

If any argument is supplied in addition to x (such as n or keep in xts::first), regardless of x's type, then xts::first is called if xts has been loaded, otherwise a helpful error.

See Also

NROW, head, tail, last

Examples

first(1:5) # [1] 1
x = data.table(x=1:5, y=6:10)
first(x) # same as x[1]

[Package data.table version 1.11.8 Index]