notin {data.table}	R Documentation

Convenience operator for checking if an example is not in a set of elements

Description

Check whether an object is absent from a table, i.e., the logical inverse of in. See examples on how missing values are being handled.

Usage

x %notin% table

Arguments

x
Vector or NULL: the values to be matched.

table
Vector or NULL: the values to be matched against.

Value

Logical vector, TRUE for each element of x absent from table, and FALSE for each element of x present in table.

See Also

match, chmatch

Examples

  11 %notin% 1:10 # TRUE
  "a" %notin% c("a", "b") # FALSE

  ## NAs on the LHS
  NA %in% 1:2
  NA %notin% 1:2
  ## NAs on the RHS
  NA %in% c(1:2,NA)
  NA %notin% c(1:2,NA)

[Package data.table version 1.15.4 Index]