subset.data.table {data.table}	R Documentation

Subsetting data.tables

Description

Returns subsets of a data.table.

Usage

  ## S3 method for class 'data.table'
subset(x, subset, select, ...)

Arguments

x
data.table to subset.

subset
logical expression indicating elements or rows to keep

select
expression indicating columns to select from data.table

...
further arguments to be passed to or from other methods

Details

The subset argument works on the rows and will be evaluated in the data.table so columns can be referred to (by name) as variables in the expression.

The data.table that is returned will maintain the original keys as long as they are not select-ed out.

Value

A data.table containing the subset of rows and columns that are selected.

See Also

subset

Examples

DT <- data.table(a=sample(c('a', 'b', 'c'), 20, replace=TRUE),
                 b=sample(c('a', 'b', 'c'), 20, replace=TRUE),
                 c=sample(20), key=c('a', 'b'))

sub <- subset(DT, a == 'a')
all.equal(key(sub), key(DT))

[Package data.table version 1.14.2 Index]