tables {data.table}	R Documentation

Display all objects of class 'data.table'

Description

Lists all data.table's in memory, including number of rows, column names and any keys.

Usage

tables(mb = TRUE, order.col = "NAME", width = 80, env=parent.frame(), silent=FALSE)

Arguments

mb
TRUE adds size of the data.table in MB to the output (slow in older versions of R).

order.col
Quoted column name to sort the output by

width
Number of characters to truncate the COLS output

env
Usually tables() is executed at the prompt where parent.frame() returns .GlobalEnv. tables() may also be useful inside functions where parent.frame() is the local scope of the function, or set it to .GlobalEnv

silent
By default tables() is expected to be called at the prompt for its compact print output. silent=TRUE prints nothing. The data statistics are returned as a data.table, silently, whether silent is TRUE or FALSE

Value

A data.table containing the information printed.

See Also

data.table, setkey, ls, objects, object.size

Examples

DT = data.table(A=1:10,B=letters[1:10])
DT2 = data.table(A=1:10000,ColB=10000:1)
setkey(DT,B)
tables()

[Package data.table version 1.9.6 Index]