Test assertions for equality, exceptions and console output

An internal testing function used in data.table test scripts that are run by test.data.table.

test(num, x, y = TRUE,
     error = NULL, warning = NULL, message = NULL,
     output = NULL, notOutput = NULL)

Arguments

num
A unique identifier for a test, helpful in identifying the source of failure when testing is not working. Currently, we use a manually-incremented system with tests formatted as n.m, where essentially n indexes an issue and m indexes aspects of that issue. For the most part, your new PR should only have one value of n (scroll to the end of inst/tests/tests.Rraw to see the next available ID) and then index the tests within your PR by increasing m. Note -- n.m is interpreted as a number, so 123.4 and 123.40 are actually the same -- please 0-pad as appropriate. Test identifiers are checked to be in increasing order at runtime to prevent duplicates being possible.

x
An input expression to be evaluated.

y
Pre-defined value to compare to x, by default TRUE.

error
When you are testing behaviour of code that you expect to fail with an error, supply the expected error message to this argument. It is interpreted as a regular expression, so you can be abbreviated, but try to include the key portion of the error so as not to accidentally include a different error message.

warning
Same as error, in the case that you expect your code to issue a warning. Note that since the code evaluates successfully, you should still supply y.

message
Same as warning but expects message exception.

output
If you are testing the printing/console output behaviour; e.g. with verbose=TRUE or options(datatable.verbose=TRUE). Again, regex-compatible and case sensitive.

notOutput
Or if you are testing that a feature does not print particular console output. Case insensitive (unlike output) so that the test does not incorrectly pass just because the string is not found due to case.

Note

NA_real_ and NaN are treated as equal, use identical if distinction is needed. See examples below.

If warning= is not supplied then you are automatically asserting no warning is expected; the test will fail if any warning does occur. Similarly for message=.

Multiple warnings are supported; supply a vector of strings to warning=. If x does not produce the correct number of warnings in the correct order, the test will fail.

Strings passed to notOutput= should be minimal; e.g. pick out single words from the output that you desire to check does not occur. The reason being so that the test does not incorrectly pass just because the output has slightly changed. For example notOutput="revised" is better than notOutput="revised flag to true". notOutput= is automatically case insensitive for this reason.

Value

Logical TRUE when test passes, FALSE when test fails. Invisibly.

See also

Examples

test = data.table:::test
test(1, x = sum(1:5), y = 15L)
test(2, log(-1), NaN, warning="NaNs")
test(3, sum("a"), error="invalid.*character")
# test failure example
stopifnot(
  test(4, TRUE, FALSE) == FALSE
)

#> Test 4 ran without errors but failed check that x equals y:
#> > x = TRUE 
#> First 1 of 1 (type 'logical'): 
#> [1] TRUE
#> > y = FALSE 
#> First 1 of 1 (type 'logical'): 
#> [1] FALSE
#> 1 element mismatch

# NA_real_ vs NaN
test(5.01, NA_real_, NaN)
test(5.03, all.equal(NaN, NA_real_))
test(5.02, identical(NaN, NA_real_), FALSE)