test.data.table {data.table}	R Documentation

Runs a set of tests.

Description

Runs a set of tests to check data.table is working correctly.

Usage

test.data.table(verbose=FALSE, pkg="pkg", silent=FALSE,
                with.other.packages=FALSE, benchmark=FALSE)

Arguments

verbose
If TRUE sets datatable.verbose to TRUE for the duration of the tests.

pkg
Root directory name under which all package content (ex: DESCRIPTION, src/, R/, inst/ etc..) resides.

silent
Logical, default FALSE, when TRUE it will not raise error on in case of test fails.

with.other.packages
Run compatibility tests with other packages.

benchmark
Run the benchmark script.

Details

Runs a series of tests. These can be used to see features and examples of usage, too. Running test.data.table will tell you the full location of the test file(s) to open.

Value

When silent equals to TRUE it will return TRUE if all tests were successful. FALSE otherwise. If silent equals to FALSE it will return TRUE if all tests were successful. Error otherwise.

See Also

data.table

Examples

## Not run: 
  library(data.table)
  test.data.table()

## End(Not run)

[Package data.table version 1.11.8 Index]