update.dev.pkg {data.table}	R Documentation

Perform update of development version of a package

Description

It will download and install package from devel repository only when new commit is available there, otherwise only PACKAGES file is transferred. Defaults are set to update data.table, other packages can be used as well. Their repository has to include git commit information in PACKAGES file.

Usage

## S3 method for class 'dev.pkg'
update(object="data.table",
       repo="https://Rdatatable.gitlab.io/data.table",
       field="Revision", type=getOption("pkgType"), lib=NULL, ...)

Arguments

object
character scalar, package name.

repo
character scalar, url of package devel repository.

field
character scalar, metadata field to use in PACKAGES file and DESCRIPTION file, default "Revision".

type
character scalar, default getOption("pkgType"), used to define if package has to be installed from sources, binaries or both.

lib
character scalar, library location where package is meant to be upgraded.

...
passed to install.packages.

Details

In case if a devel repository does not provide binaries user will need development tools installed for package compilation, like Rtools on Windows, and eventually set type="source".

Value

NULL.

Note

Package namespace is unloaded before attempting to install newer version.

See Also

data.table

Examples

  # data.table::update.dev.pkg()

[Package data.table version 1.14.2 Index]