last {data.table}	R Documentation

First/last item of an object

Description

Returns the first/last item of a vector or list, or the first/last row of a data.frame or data.table. The main difference to head/tail is that the default for n is 1 rather than 6.

Usage

first(x, n=1L, ...)
last(x, n=1L, ...)

Arguments

x
A vector, list, data.frame or data.table. Otherwise the S3 method of xts::first is deployed.

n
A numeric vector length 1. How many items to select.

...
Not applicable for data.table first/last. Any arguments here are passed through to xts's first/last.

Value

If no other arguments are supplied it depends on the type of x. The first/last item of a vector or list. The first/last row of a data.frame or data.table. For other types, or if any argument is supplied in addition to x (such as n, or keep in xts) regardless of x's type, then xts::first/ xts::last is called if xts has been loaded, otherwise utils::head/utils::tail.

See Also

NROW, head, tail

Examples

first(1:5) # [1] 1
x = data.table(x=1:5, y=6:10)
first(x) # same as head(x, 1)

last(1:5) # [1] 5
x = data.table(x=1:5, y=6:10)
last(x) # same as tail(x, 1)

[Package data.table version 1.14.2 Index]