test.data.table {data.table}	R Documentation

Runs a set of tests.

Description

Runs a set of tests to check data.table is working correctly.

Usage

test.data.table(verbose=FALSE, pkg="pkg")

Arguments

verbose
If TRUE sets datatable.verbose to TRUE for the duration of the tests.

pkg
Root directory name under which all package content (ex: DESCRIPTION, src/, R/, inst/ etc..) resides.

Details

Runs a series of tests. These can be used to see features and examples of usage, too. Running test.data.table will tell you the full location of the test file(s) to open.

Value

TRUE if all tests were successful. FALSE otherwise.

See Also

data.table

Examples

## Not run: 
test.data.table()

## End(Not run)

[Package data.table version 1.9.6 Index]